requirements.txt
pandas>=2.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-vertexai>=0.0.1
langchain-experimental>=0.0.1
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
openpyxl>=3.1.0
google-cloud-aiplatform>=1.38.0
tabulate
streamlit
